updates with config
updated icons and read api key from config
