+++ .gitignore
@@ -48,3 +48,5 @@ app.*.map.json
 ios/Podfile.lock
 
 pubspec.lock
+
+assets/secrets.json
