--- assets/map.html
@@ -1,31 +1,0 @@
-<!-- Include Leaflet CSS and JS from CDN -->
-<link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" />
-<script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
-
-<div id="map" style="width:100%; height:100%;"></div>
-<script>
-    // Initialize map at a given center and zoom
-    var map = L.map('map').setView([50.08, 14.42], 13);
-    // Mapy.cz base layer (terrain map)
-    const baseLayer = L.tileLayer(`https://api.mapy.cz/v1/maptiles/basic/256/{z}/{x}/{y}?apikey=YOUR_API_KEY`, {
-      maxZoom: 19,
-      attribution: '<a href="https://api.mapy.cz/copyright" target="_blank">&copy; Seznam.cz a.s.</a>'
-    }).addTo(map);
-    // Aerial layer (satellite imagery) and overlay for labels (names)
-    const aerialLayer = L.tileLayer(`https://api.mapy.cz/v1/maptiles/aerial/256/{z}/{x}/{y}?apikey=YOUR_API_KEY`, { maxZoom: 20 });
-    const labelsLayer = L.tileLayer(`https://api.mapy.cz/v1/maptiles/names-overlay/256/{z}/{x}/{y}?apikey=YOUR_API_KEY`, { maxZoom: 20 });
-    // Group aerial + labels for a hybrid map
-    const hybridLayer = L.layerGroup([aerialLayer, labelsLayer]);
-    // Layer control to switch between base and hybrid (satellite)
-    L.control.layers({ "Map": baseLayer, "Satellite": hybridLayer }).addTo(map);
-    // (Optional) Add a custom grid overlay (example: draw a red polyline grid)
-    const latLines = [];
-    for (let lat = 50.0; lat <= 50.5; lat += 0.1) {
-      latLines.push(L.polyline([[lat, 14.3], [lat, 14.7]], {color: 'red'}));
-    }
-    const lngLines = [];
-    for (let lng = 14.3; lng <= 14.7; lng += 0.1) {
-      lngLines.push(L.polyline([[50.0, lng], [50.5, lng]], {color: 'red'}));
-    }
-    L.layerGroup([...latLines, ...lngLines]).addTo(map);
-</script>
